# IKM-WH/backend-node.js/swagger.yaml
openapi: 3.0.3
info:
  title: Festival Management System API
  description: |
    Полнофункциональный API для системы управления музыкальными фестивалями.
    
    ## Возможности
    - Управление заявками на участие в фестивале
    - Администрирование групп, мероприятий и фестивалей
    - Управление мерчендайзом и коллекционными карточками
    - Система карточек участников с загрузкой изображений
    - Управление афишами и постерами
    - Контент-менеджмент для сайта
    - Статистика и аналитика
    
    ## Аутентификация
    API использует JWT токены для аутентификации администраторов.
    Публичные эндпоинты доступны без аутентификации.
    
  version: 2.0.0
  contact:
    name: Festival Management System
    email: support@festival-system.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.festival-system.com/api
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен для аутентификации администраторов
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          description: Текущая страница
        limit:
          type: integer
          description: Количество элементов на странице
        total:
          type: integer
          description: Общее количество элементов
        totalPages:
          type: integer
          description: Общее количество страниц
...
    FestivalApplication:
      type: object
      properties:
        id:
          type: string
          description: Уникальный идентификатор заявки
        groupName:
          type: string
          description: Название группы или исполнителя
        artistName:
          type: string
          description: Имя артиста (для сольных исполнителей)
        bandName:
          type: string
          description: Название группы
        contactTelegram:
          type: string
          description: Telegram контакт
        contactPhone:
          type: string
          description: Номер телефона
        description:
          type: string
          description: Описание группы/исполнителя
        socialLinks:
          $ref: '#/components/schemas/SocialLinks'
        genre:
          type: string
          description: Музыкальный жанр
        members:
          type: array
          items:
            type: string
          description: Участники группы
        status:
          type: string
          enum: [pending, approved, rejected]
          description: Статус заявки
        submittedAt:
          type: string
          format: date-time
          description: Дата подачи заявки
        reviewedAt:
          type: string
          format: date-time
          description: Дата рассмотрения заявки
        reviewedBy:
          type: string
          description: Кто рассмотрел заявку
        notes:
          type: string
          description: Заметки администратора
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateApplicationRequest:
      type: object
      required:
        - groupName
        - contactTelegram
        - contactPhone
      properties:
        groupName:
          type: string
          minLength: 2
          description: Название группы или исполнителя
        artistName:
          type: string
          description: Имя артиста
        bandName:
          type: string
          description: Название группы
        contactTelegram:
          type: string
          description: Telegram контакт
        contactPhone:
          type: string
          description: Номер телефона
        description:
          type: string
          description: Описание группы/исполнителя
        socialLinks:
          $ref: '#/components/schemas/SocialLinks'
        genre:
          type: string
          description: Музыкальный жанр
        members:
          type: array
          items:
            type: string
          description: Участники группы
...
    Group:
      type: object
      properties:
        id:
          type: string
          description: Уникальный идентификатор группы
        name:
          type: string
          description: Название группы
        genre:
          type: string
          description: Музыкальный жанр
        description:
          type: string
          description: Описание группы
        members:
          type: array
          items:
            type: string
          description: Участники группы
        image:
          type: string
          description: URL изображения группы
        vkLink:
          type: string
          description: Ссылка на VK
        tgLink:
          type: string
          description: Ссылка на Telegram
        socialLinks:
          $ref: '#/components/schemas/SocialLinks'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Описание ошибки
        details:
          type: object
          description: Дополнительные детали ошибки
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Validation failed"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Поле с ошибкой
              message:
                type: string
                description: Сообщение об ошибке
paths:
...
  /applications:
    get:
      tags: [Applications]
      summary: Получение списка заявок
      description: Получение всех заявок на участие (только для администраторов)
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          description: Фильтр по статусу заявки
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: page
          in: query
          description: Номер страницы
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Количество элементов на странице
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        200:
          description: Список заявок
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FestivalApplication'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        401:
          description: Неавторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          description: Недостаточно прав доступа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [Applications]
      summary: Подача заявки на участие
      description: Создание новой заявки на участие в фестивале (публичный эндпоинт)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicationRequest'
            example:
              groupName: "The Rock Stars"
              contactTelegram: "@rockstars"
              contactPhone: "+7900123456"
              genre: "Rock"
              description: "Мы молодая рок-группа из Москвы"
              members: ["Иван Петров", "Мария Сидорова", "Алексей Козлов"]
              socialLinks:
                instagram: "rockstars_band"
                youtube: "RockStarsChannel"
      responses:
        201:
          description: Заявка успешно создана
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/FestivalApplication'
        400:
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
...
  /groups:
    get:
      tags: [Groups]
      summary: Получение списка групп
      description: Получение всех музыкальных групп
      parameters:
        - name: genre
          in: query
          description: Фильтр по жанру
          schema:
            type: string
        - name: page
          in: query
          description: Номер страницы
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Количество элементов на странице
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        200:
          description: Список групп
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Group'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags: [Groups]
      summary: Создание новой группы
      description: Добавление новой музыкальной группы (только для администраторов)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                genre:
                  type: string
                description:
                  type: string
                members:
                  type: array
                  items:
                    type: string
                image:
                  type: string
                  format: binary
                  description: Изображение группы
      responses:
        201:
          description: Группа успешно создана
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Group'
        400:
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
...
  /events:
    get:
      tags: [Events]
      summary: Получение списка событий
      description: Получение всех событий фестиваля
      parameters:
        - name: status
          in: query
          description: Фильтр по статусу события
          schema:
            type: string
            enum: [upcoming, ongoing, completed, cancelled]
        - name: upcoming
          in: query
          description: Только предстоящие события
          schema:
            type: boolean
        - name: page
          in: query
          description: Номер страницы
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Количество элементов на странице
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        200:
          description: Список событий
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags: [Events]
      summary: Создание нового события
      description: Добавление нового события фестиваля (только для администраторов)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                date:
                  type: string
                  format: date
                time:
                  type: string
                venue:
                  type: string
                address:
                  type: string
                performers:
                  type: array
                  items:
                    type: string
                groupId:
                  type: string
                image:
                  type: string
                  format: binary
                  description: Изображение события
      responses:
        201:
          description: Событие успешно создано
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Event'
...
  /festivals:
    get:
      tags: [Festivals]
      summary: Получение списка фестивалей
      description: Получение всех фестивалей
      parameters:
        - name: status
          in: query
          description: Фильтр по статусу фестиваля
          schema:
            type: string
            enum: [upcoming, ongoing, completed, cancelled]
        - name: page
          in: query
          description: Номер страницы
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Количество элементов на странице
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        200:
          description: Список фестивалей
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Festival'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags: [Festivals]
      summary: Создание нового фестиваля
      description: Добавление нового фестиваля (только для администраторов)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFestivalRequest'
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                date:
                  type: string
                  format: date
                venue:
                  type: string
                description:
                  type: string
                time:
                  type: string
                address:
                  type: string
                performers:
                  type: array
                  items:
                    type: string
                image:
                  type: string
                  format: binary
                  description: Изображение фестиваля
      responses:
        201:
          description: Фестиваль успешно создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Festival'
...
  /merch:
    get:
      tags: [Merchandise]
      summary: Получение каталога товаров
      description: Получение всех товаров мерчендайза
      parameters:
        - name: category
          in: query
          description: Фильтр по категории товара
          schema:
            type: string
        - name: inStock
          in: query
          description: Только товары в наличии
          schema:
            type: boolean
        - name: page
          in: query
          description: Номер страницы
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Количество элементов на странице
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        200:
          description: Каталог товаров
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MerchItem'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags: [Merchandise]
      summary: Добавление нового товара
      description: Создание нового товара мерчендайза (только для администраторов)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMerchItemRequest'
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                price:
                  type: number
                stock:
                  type: integer
                category:
                  type: string
                image:
                  type: string
                  format: binary
                  description: Изображение товара
      responses:
        201:
          description: Товар успешно создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/MerchItem'
...
  /cards:
    get:
      tags: [Cards]
      summary: Получение наборов карточек
      description: Получение всех наборов коллекционных карточек
      parameters:
        - name: festivalId
          in: query
          description: Фильтр по ID фестиваля
          schema:
            type: string
        - name: available
          in: query
          description: Только доступные наборы
          schema:
            type: boolean
        - name: page
          in: query
          description: Номер страницы
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Количество элементов на странице
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        200:
          description: Список наборов карточек
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CardSet'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags: [Cards]
      summary: Создание набора карточек
      description: Создание нового набора коллекционных карточек (только для администраторов)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardSetRequest'
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                festivalId:
                  type: string
                cardsInSet:
                  type: integer
                setsAvailable:
                  type: integer
                price:
                  type: number
                image:
                  type: string
                  format: binary
                  description: Изображение набора
      responses:
        201:
          description: Набор карточек успешно создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/CardSet'
...
  /participant-cards:
    get:
      tags: [Participant Cards]
      summary: Получение карточек участников
      description: Получение всех карточек участников фестиваля
      parameters:
        - name: festivalId
          in: query
          description: Фильтр по ID фестиваля
          schema:
            type: string
        - name: groupName
          in: query
          description: Фильтр по названию группы
          schema:
            type: string
        - name: active
          in: query
          description: Только активные карточки
          schema:
            type: boolean
        - name: page
          in: query
          description: Номер страницы
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Количество элементов на странице
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        200:
          description: Список карточек участников
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ParticipantCard'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags: [Participant Cards]
      summary: Создание карточек участников
      description: Создание новых карточек участников с загрузкой изображений (только для администраторов)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - festivalId
                - groupName
                - eventId
                - cards
              properties:
                festivalId:
                  type: string
                  description: ID фестиваля
                groupName:
                  type: string
                  description: Название группы
                eventId:
                  type: string
                  description: ID события
                cards:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Файлы карточек (изображения)
                price:
                  type: number
                  description: Цена набора карточек
                setsAvailable:
                  type: integer
                  description: Доступное количество наборов
                isActive:
                  type: boolean
                  description: Активность набора
      responses:
        201:
          description: Карточки участников успешно созданы
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ParticipantCard'
...
  /posters:
    get:
      tags: [Posters]
      summary: Получение списка постеров
      description: Получение всех постеров и афиш
      parameters:
        - name: active
          in: query
          description: Только активные постеры
          schema:
            type: boolean
        - name: page
          in: query
          description: Номер страницы
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Количество элементов на странице
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        200:
          description: Список постеров
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Poster'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags: [Posters]
      summary: Загрузка нового постера
      description: Создание нового постера с загрузкой изображения (только для администраторов)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
                - image
              properties:
                title:
                  type: string
                  description: Название постера
                description:
                  type: string
                  description: Описание постера
                eventDate:
                  type: string
                  format: date
                  description: Дата связанного события
                isActive:
                  type: boolean
                  description: Активность постера
                image:
                  type: string
                  format: binary
                  description: Изображение постера
      responses:
        201:
          description: Постер успешно создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Poster'
...
  /content:
    get:
      tags: [Content]
      summary: Получение контента сайта
      description: Получение всего контента для сайта
      parameters:
        - name: type
          in: query
          description: Фильтр по типу контента
          schema:
            type: string
            enum: [text, image, list]
        - name: key
          in: query
          description: Поиск по ключу контента
          schema:
            type: string
      responses:
        200:
          description: Контент сайта
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FestivalContent'

    post:
      tags: [Content]
      summary: Создание контента
      description: Создание нового контента для сайта (только для администраторов)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContentRequest'
      responses:
        201:
          description: Контент успешно создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/FestivalContent'
...
  /stats:
    get:
      tags: [Statistics]
      summary: Общая статистика системы
      description: Получение общей статистики платформы (только для администраторов)
      security:
        - BearerAuth: []
      responses:
        200:
          description: Статистика системы
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Statistics'
        401:
          description: Неавторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
security:
  - BearerAuth: []

# Examples and additional documentation
...